requirements.txt
pika
redis
PyYAML
